requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
openai
pdfplumber
pytesseract
Pillow
PyMuPDF
python-docx
pandas
google-generativeai
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
boto3
reportlab
bcrypt
python-dotenv
markdown
beautifulsoup4
